+++ scripts/macros/2024/mechanics/heroicInspiration.js
@@ -4,7 +4,8 @@ async function attack(workflow) {
     let selection = await dialogUtils.selectDie([workflow.attackRoll], 'CHRISPREMADES.HeroicInspiration.Name', genericUtils.format('CHRISPREMADES.Dialog.Use', {itemName: genericUtils.translate('CHRISPREMADES.HeroicInspiration.Name')}), {buttons: 'yesNo'});
     if (!selection) return;
     let positions = selection[0].split('-').map(i => Number(i));
-    let roll = await rollUtils.rollDice('1d' + workflow.attackRoll.terms[positions[1]].faces, {chatMessage: true, flavor: genericUtils.translate('CHRISPREMADES.HeroicInspiration.Name')});
+    let mode = game.settings.get('core', 'rollMode') || 'publicroll';
+    let roll = await rollUtils.rollDice('1d' + workflow.attackRoll.terms[positions[1]].faces, {chatMessage: true, mode, flavor: genericUtils.translate('CHRISPREMADES.HeroicInspiration.Name')});
     let newRoll = rollUtils.updateDieResult(workflow.attackRoll, positions[1], positions[2], roll.roll.total);
     await workflow.setAttackRoll(newRoll);
     await genericUtils.update(workflow.actor, {'system.attributes.inspiration': false});
@@ -14,7 +15,8 @@ async function damage(workflow) {
     let selection = await dialogUtils.selectDie(workflow.damageRolls, 'CHRISPREMADES.HeroicInspiration.Name', genericUtils.format('CHRISPREMADES.Dialog.Use', {itemName: genericUtils.translate('CHRISPREMADES.HeroicInspiration.Name')}), {buttons: 'yesNo'});
     if (!selection) return;
     let positions = selection[0].split('-').map(i => Number(i));
-    let roll = await rollUtils.rollDice('1d' + workflow.damageRolls[positions[0]].terms[positions[1]].faces, {chatMessage: true, flavor: genericUtils.translate('CHRISPREMADES.HeroicInspiration.Name')});
+    let mode = game.settings.get('core', 'rollMode') || 'publicroll';
+    let roll = await rollUtils.rollDice('1d' + workflow.damageRolls[positions[0]].terms[positions[1]].faces, {chatMessage: true, mode, flavor: genericUtils.translate('CHRISPREMADES.HeroicInspiration.Name')});
     let newRoll = rollUtils.updateDieResult(workflow.damageRolls[positions[0]], positions[1], positions[2], roll.roll.total);
     workflow.damageRolls[[positions[0]]] = newRoll;
     await workflow.setDamageRolls(workflow.damageRolls);
@@ -25,7 +27,8 @@ async function saveSkillCheck(roll, actor) {
     let selection = await dialogUtils.selectDie([roll], 'CHRISPREMADES.HeroicInspiration.Name', genericUtils.format('CHRISPREMADES.Dialog.Use', {itemName: genericUtils.translate('CHRISPREMADES.HeroicInspiration.Name')}), {buttons: 'yesNo'});
     if (!selection) return;
     let positions = selection[0].split('-').map(i => Number(i));
-    let rolled = await rollUtils.rollDice('1d' + roll.terms[positions[1]].faces, {chatMessage: true, flavor: genericUtils.translate('CHRISPREMADES.HeroicInspiration.Name')});
+    let mode = game.settings.get('core', 'rollMode') || 'publicroll';
+    let rolled = await rollUtils.rollDice('1d' + roll.terms[positions[1]].faces, {chatMessage: true, mode, flavor: genericUtils.translate('CHRISPREMADES.HeroicInspiration.Name')});
     let newRoll = rollUtils.updateDieResult(roll, positions[1], positions[2], rolled.roll.total);
     await genericUtils.update(actor, {'system.attributes.inspiration': false});
     return newRoll;